evaluation package refactor
