+++ src/main/java/com/configcat/ConfigCatHooks.java
@@ -1,5 +1,7 @@
 package com.configcat;
 
+import com.configcat.evaluation.EvaluationDetails;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
