+++ src/main/java/com/configcat/Constants.java
@@ -1,12 +1,11 @@
 package com.configcat;
 
-final class Constants {
+public final class Constants {
     private Constants() { /* prevent from instantiation*/ }
 
-    static final long DISTANT_FUTURE = Long.MAX_VALUE;
-    static final long DISTANT_PAST = 0;
-    static final String CONFIG_JSON_NAME = "config_v5";
-    static final String SERIALIZATION_FORMAT_VERSION = "v1";
-
-    static final String VERSION = "8.1.0";
+    public static final long DISTANT_FUTURE = Long.MAX_VALUE;
+    public static final long DISTANT_PAST = 0;
+    public static final String CONFIG_JSON_NAME = "config_v5";
+    public static  final String SERIALIZATION_FORMAT_VERSION = "v1";
+    public static final String VERSION = "8.1.0";
 }
